Move `vaultsecrets` resources to their respective folder
Also fixes the CODEOWNER for `vaultsecrets` by moving them to the correct location
