+++ internal/provider/vaultsecrets/data_source_vault_secrets_app_test.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package provider
+package vaultsecrets_test
 
 import (
 	"context"
@@ -14,6 +14,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 	"github.com/hashicorp/terraform-provider-hcp/internal/clients"
+	"github.com/hashicorp/terraform-provider-hcp/internal/provider/acctest"
 )
 
 func TestAcc_dataSourceVaultSecretsAppMigration(t *testing.T) {
@@ -26,7 +27,7 @@ func TestAcc_dataSourceVaultSecretsAppMigration(t *testing.T) {
 	secondSecretValue := "hey, this is version 2!"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck: func() { testPreCheck(t) },
+		PreCheck: func() { acctest.PreCheck(t) },
 		Steps: []resource.TestStep{
 			// Create two secrets, one with an additional version and check the latest secrets from data source
 			{
@@ -55,7 +56,7 @@ func TestAcc_dataSourceVaultSecretsAppMigration(t *testing.T) {
 				),
 			},
 			{
-				ProtoV5ProviderFactories: testProtoV5ProviderFactories,
+				ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
 				Config: fmt.Sprintf(`
 				data "hcp_vault_secrets_app" "example" {
 					app_name    = %q
@@ -82,8 +83,8 @@ func TestAcc_dataSourceVaultSecretsApp(t *testing.T) {
 	secondSecretValue := "hey, this is version 2!"
 
 	resource.Test(t, resource.TestCase{
-		PreCheck:                 func() { testPreCheck(t) },
-		ProtoV5ProviderFactories: testProtoV5ProviderFactories,
+		PreCheck:                 func() { acctest.PreCheck(t) },
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
 		Steps: []resource.TestStep{
 			// Create two secrets, one with an additional version and check the latest secrets from data source
 			{
@@ -118,18 +119,14 @@ func TestAcc_dataSourceVaultSecretsApp(t *testing.T) {
 func createTestApp(t *testing.T, appName string) {
 	t.Helper()
 
-	client, err := newDefaultClient()
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
+	client := acctest.HCPClients(t)
 
 	loc := &sharedmodels.HashicorpCloudLocationLocation{
 		OrganizationID: client.Config.OrganizationID,
 		ProjectID:      client.Config.ProjectID,
 	}
 
-	_, err = clients.CreateVaultSecretsApp(context.Background(), client, loc, appName, "app description")
+	_, err := clients.CreateVaultSecretsApp(context.Background(), client, loc, appName, "app description")
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -138,17 +135,14 @@ func createTestApp(t *testing.T, appName string) {
 func createTestAppSecret(t *testing.T, appName, secretName, secretValue string) {
 	t.Helper()
 
-	client, err := newDefaultClient()
-	if err != nil {
-		t.Fatal(err)
-	}
+	client := acctest.HCPClients(t)
 
 	loc := &sharedmodels.HashicorpCloudLocationLocation{
 		OrganizationID: client.Config.OrganizationID,
 		ProjectID:      client.Config.ProjectID,
 	}
 
-	_, err = clients.CreateVaultSecretsAppSecret(context.Background(), client, loc, appName, secretName, secretValue)
+	_, err := clients.CreateVaultSecretsAppSecret(context.Background(), client, loc, appName, secretName, secretValue)
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -157,18 +151,14 @@ func createTestAppSecret(t *testing.T, appName, secretName, secretValue string) 
 func deleteTestAppSecret(t *testing.T, appName, secretName string) {
 	t.Helper()
 
-	client, err := newDefaultClient()
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
+	client := acctest.HCPClients(t)
 
 	loc := &sharedmodels.HashicorpCloudLocationLocation{
 		OrganizationID: client.Config.OrganizationID,
 		ProjectID:      client.Config.ProjectID,
 	}
 
-	err = clients.DeleteVaultSecretsAppSecret(context.Background(), client, loc, appName, secretName)
+	err := clients.DeleteVaultSecretsAppSecret(context.Background(), client, loc, appName, secretName)
 	if err != nil {
 		t.Error(err)
 	}
@@ -177,18 +167,13 @@ func deleteTestAppSecret(t *testing.T, appName, secretName string) {
 func deleteTestApp(t *testing.T, appName string) {
 	t.Helper()
 
-	client, err := newDefaultClient()
-	if err != nil {
-		t.Fatal(err)
-		return
-	}
-
+	client := acctest.HCPClients(t)
 	loc := &sharedmodels.HashicorpCloudLocationLocation{
 		OrganizationID: client.Config.OrganizationID,
 		ProjectID:      client.Config.ProjectID,
 	}
 
-	err = clients.DeleteVaultSecretsApp(context.Background(), client, loc, appName)
+	err := clients.DeleteVaultSecretsApp(context.Background(), client, loc, appName)
 	if err != nil {
 		t.Error(err)
 	}
