+++ internal/provider/vaultsecrets/data_source_vault_secrets_secret.go
@@ -1,7 +1,7 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package provider
+package vaultsecrets
 
 import (
 	"context"
