+++ internal/provider/vaultsecrets/resource_vault_secrets_secret_test.go
@@ -1,19 +1,20 @@
 // Copyright (c) HashiCorp, Inc.
 // SPDX-License-Identifier: MPL-2.0
 
-package provider
+package vaultsecrets_test
 
 import (
 	"fmt"
 	"testing"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-provider-hcp/internal/provider/acctest"
 )
 
 func TestAccVaultSecretsResourceSecret(t *testing.T) {
 	testAppName := generateRandomSlug()
 	resource.Test(t, resource.TestCase{
-		ProtoV5ProviderFactories: testProtoV5ProviderFactories,
+		ProtoV5ProviderFactories: acctest.ProtoV5ProviderFactories,
 		Steps: []resource.TestStep{
 			{
 				PreConfig: func() {
